Moved Uberblock out of zfs.rs
